// src/types.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { NotifyLevel } from "../src-tauri/bindings/NotifyLevel";
import type { NotifyRange } from "../src-tauri/bindings/NotifyRange";

/**
 * アプリ全体のユーザー設定
 *
 * * デフォルト値で生成する機能
 * * OSに合わせた場所へ、JSONで読み書きする機能
 * * 設定値のバリデーションチェック機能
 */
export type Config = {
  sourcePath: string;
  destinationPath: string;
  recursive: boolean;
  allFilesEnabled: boolean;
  extensions: Array<string>;
  isNotify: boolean;
  isNotifyUnsaved: boolean;
  notifyInterval: number;
  isShown: boolean;
  autoStart: boolean;
};

/**
 * UIへの送信用のデータ型
 */
export type NotifyPayload = {
  level: NotifyLevel;
  range: NotifyRange;
  title: string;
  body: string;
};